Вынес стандартные действия по получению сущностей в отдельный трейт
