+++ Services/Traits/BitrixEntity.php
@@ -5,7 +5,7 @@ namespace Services\Traits;
 
 use Bitrix\Main\Data\Cache;
 
-trait BitrixEntityTrait
+trait BitrixEntity
 {
     private $iblockId;
 
@@ -92,32 +92,6 @@ trait BitrixEntityTrait
             $result[$item["CODE"] ?? $item["ID"]] = $item;
         }
         return $result;
-    }
-
-    public function getEntityById($iblockId, $id, $select = [])
-    {
-        $result = $this->setIblock($iblockId)
-            ->setFilter(["ID" => $id])
-            ->setSelect($select)
-            ->getListFromDb();
-        return array_shift($result);
-    }
-
-    public function getEntityByCode($iblockId, $code, $select = [])
-    {
-        $result = $this->setIblock($iblockId)
-            ->setFilter(["CODE" => $code])
-            ->setSelect($select)
-            ->getListFromDb();
-        return array_shift($result);
-    }
-
-    public function getEntitiesByFilter($iblockId, $filter, $select = [])
-    {
-        return $this->setIblock($iblockId)
-            ->setFilter($filter)
-            ->setSelect($select)
-            ->getListFromDb();
     }
 
 }
